requirements.txt
fastapi~=0.110.0
pydantic~=2.6.3
uvicorn~=0.27.1
python-dotenv~=1.0.1
fastapi-users~=12.1.3

alembic~=1.12.1
sqlalchemy~=2.0.28
psycopg~=3.1.18
asyncpg~=0.29.0